refactor(gallery): split page.tsx, 7517 lines holding fifteen unrelated components
GalleryInner is ~4,000 of those lines and stays where it is — untangling its
state is a separate job with real risk. What moved is everything that was
merely parked behind it:

  _shared.tsx            types, pure helpers, the module caches       240
  _gallery-css.ts        268 lines of selectors                       271
  _components/dropdowns  pickers, dropdowns, provider icons           939
  _components/GalleryCard                                             415
  _components/Lightbox                                                375
  _components/highlight  JSON/YAML syntax colouring                   139
  _components/DownloadToast                                           105
  _components/empty-states                                            136

page.tsx: 7517 -> 4974.

This is a move, not a rewrite. Verified by diffing the line multiset of the
original against every new file: the only lines that differ are the five
touching _galleryDragItem, which had to become a holder object — an imported
`let` cannot be reassigned, and GalleryCard writes it on drag start.

EmptyState went with the dead code; it was defined and never rendered, which
only became visible once it stopped being buried.

Checked in a real browser, not just at the type level: three GalleryCard tiles
render, clicking one opens the Lightbox with its prompt drawn by the
renderLightboxPrompt that moved to _shared, and the console is clean. The
import graph is acyclic — every new module depends only on _shared.

tsc clean, build clean, 7/7 tests, eslint errors unchanged at 74 and warnings
down from 87 to 84.
